+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0
+
+- sdk version lowered to 2.15.0
+- internet_connection_checker_plus package upgraded to 2.7.0
+- removed google fonts
+- TextStyle property added
+
 ## 0.0.9
 
 - sdk version lowered to 2.17.0